+++ fbw-a380x/src/systems/instruments/src/MFD/pages/FMS/F-PLN/MfdFmsFplnDirectTo.tsx
@@ -167,7 +167,6 @@ export class MfdFmsFplnDirectTo extends FmsPage<MfdFmsFplnDirectToProps> {
                     idPrefix={`${this.props.mfd.uiService.captOrFo}_MFD_directToDropdown`}
                     selectedIndex={this.selectedWaypointIndex}
                     values={this.availableWaypoints}
-                    freeTextAllowed
                     containerStyle="width: 175px;"
                     alignLabels="flex-start"
                     onModified={(i, text) => {