+++ .github/CHANGELOG.md
@@ -63,6 +63,7 @@
 1. [A32NX/FMS] Change GA PERF speeds to match APPR PERF speeds - @Jonny23787 (Jonathan)
 1. [EFB] Fixed overflow on settings pages when the page is too long - @heclak (Heclak)
 1. [EFB] Renamed "Weight Unit" pin program to "US Units" to reflect it's actual effect and real-world name - @tracernz (Mike)
+1. [A380X/MFD] Fix formatting issue on inputField in dropDownMenu - @Lucas-IQ21 (Lucas)
 
 ## 0.13.0